Update myload
automysqlbackup can/will add dates and other "_" separated information to the file name. Added awk to parse the first parameter from the "_" separated filename to be used as the database name. In the event there are no underscores in the filename, awk will simply return the string that was passed to it.